[ADD] motor code to test in main

+++ 22.WRITE_ESC/src/mcu/pwm/sensor_interface/pwm_write.h
@@ -2,4 +2,4 @@
 
 #include "stm32f0xx_tim.h"
 
-void set_duty_cycle(uint16_t TIM_Channel, uint8_t percentage);
+void set_duty_cycle(uint16_t TIM_Channel, uint16_t percentage);
